requirements.txt
pyaudio>=0.2.13
wave>=0.0.2
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
keyboard>=0.13.5
playsound>=1.3.0
